+++ src/offkai_bot/alerts/task.py
@@ -6,9 +6,7 @@ from dataclasses import dataclass, field
 import discord
 
 from offkai_bot.errors import BotCommandError  # Import base error for catching known issues
-
-# Assuming perform_close_event is correctly placed and importable
-# Adjust the import path if necessary based on your project structure
+from offkai_bot.event_actions import perform_close_event
 
 _log = logging.getLogger(__name__)
 
@@ -79,8 +77,6 @@ class CloseOffkaiTask(Task):
         try:
             # Call the core closing logic function
             # self.client is inherited from the Task dataclass
-
-            from offkai_bot.event_actions import perform_close_event
 
             await perform_close_event(self.client, self.event_name, self.close_msg)
             _log.info(f"Successfully executed automatic closure for event: '{self.event_name}'")
