+++ src/offkai_bot/data/event.py
@@ -8,6 +8,7 @@ from datetime import UTC, datetime
 # Use relative imports for sibling modules within the package
 from offkai_bot.config import get_config
 from offkai_bot.data.encoders import DataclassJSONEncoder
+from offkai_bot.data.response import get_responses, get_waitlist
 from offkai_bot.errors import (
     CapacityReductionError,
     CapacityReductionWithWaitlistError,
@@ -435,9 +436,6 @@ def update_event_details(
 
     # Validate max_capacity changes
     if max_capacity is not None and event.max_capacity != max_capacity:
-        # Import here to avoid circular imports
-        from offkai_bot.data.response import get_responses, get_waitlist
-
         # Get current attendee count and waitlist
         responses = get_responses(event_name)
         current_count = sum(1 + r.extra_people for r in responses)
@@ -534,8 +532,6 @@ def set_event_open_status(event_name: str, target_open_status: bool) -> Event:
         _log.info(f"Event '{event_name}' reopened, cleared closed_attendance_count.")
     else:
         # Closing the event - capture current attendance count
-        from offkai_bot.data.response import get_responses
-
         responses = get_responses(event_name)
         event.closed_attendance_count = sum(1 + r.extra_people for r in responses)
         _log.info(f"Event '{event_name}' closed with {event.closed_attendance_count} attendees.")
